Moved 'roles' outside of representation
