+++ mupt/interfaces/mdanalysis/strategies.py
@@ -14,7 +14,7 @@ import numpy as np
 from ...chemistry.core import BOND_ORDER
 from ...mupr.embedding import ConnectorReference
 from ...mupr.primitives import Primitive
-from ...mupr.roles import PrimitiveRole
+from ...roles import PrimitiveRole
 
 def _pdb_resname(label: Hashable, resname_map: dict[str, str]) -> str:
     """Map a residue label to a PDB-compliant 3-character residue name.
